+++ redberry/models/migrate.py
@@ -1,5 +1,6 @@
 import logging
 from sqlalchemy.sql.elements import quoted_name
+import importlib
 
 from redberry.blueprint import cms
 from redberry.models import RedPost, RedCategory, RedVersion
@@ -13,7 +14,6 @@ class RedMigrator:
         self.logger = logging.getLogger('redberry')
 
     def do_migration(self):
-        self.logger.warning("--- Migrations %s" % RedVersion.all_migrations())
         for model in [RedVersion, RedPost, RedCategory]:
             if not self.db.engine.has_table(quoted_name(model.__table__, True)):
                 self.logger.warning("Creating database table %s" % model.__tablename__)
@@ -22,6 +22,15 @@ class RedMigrator:
         if not self.db.engine.has_table(categories_posts.name):
             self.logger.warning("Creating database join table %s" % categories_posts.name)
             categories_posts.create(self.db.engine)
+
+        # Run migrations
+        migrations = RedVersion.all_migrations()
+        self.logger.warning("--- Migrations %s" % migrations)
+
+        # Migrations ['201612311100_add_publish_state.py', '201612311300_dummy.py']
+        for migration in migrations:
+            migration_file = migration.split(".")[0]  # eg "201612311100_add_publish_state"
+            self.run_migration(migration_file=migration_file)
 
     def initialize_samples(self):
         if not RedCategory.query.count():
@@ -36,3 +45,26 @@ class RedMigrator:
             self.db.session.add(post)
 
         self.db.session.flush()
+
+    def run_migration(self, migration_file, method='upgrade'):
+
+        if RedVersion.already_run(migration_file):
+            self.logger.warn("Migration %s has already been run, skipping." % migration_file)
+            return
+
+        try:
+            module = importlib.import_module('redberry.models.migrations.%s' % migration_file)
+
+            if method == 'upgrade':
+                self.logger.info("Running upgrade on %s" % migration_file)
+                module.upgrade(self.db)
+            else:
+                self.logger.info("Running downgrade on %s" % migration_file)
+                module.downgrade(self.db)
+
+            RedVersion.store_migration(migration_file)
+
+        except Exception, e:
+            self.logger.error("Error running %s" % migration_file)
+            self.logger.error(e)
+